feat: 修改Byte类型属性为Integer

+++ polaris-api/src/main/java/com/achobeta/api/dto/user/QueryUserInfoResponseDTO.java
@@ -30,7 +30,7 @@ public class QueryUserInfoResponseDTO implements Serializable {
     private String phone;
 
     @FieldDesc(name = "用户性别")
-    private Byte gender;
+    private Integer gender;
 
     @FieldDesc(name = "身份证号")
     private String idCard;
